+++ NEWS.rst
@@ -11,6 +11,7 @@ Changed
 
 - mir.orbis now uses Fire for commands.
   - `python -m mir.orbis.cmd.hash` is now `python -m mir.orbis hash`.
+- `hash` command is now `index`.
 - `pictus` module renamed to `indexing`.
 
 0.6.0 (2018-01-01)